# docs/api/openapi-query.yaml
openapi: 3.1.0
info:
  title: loglake query
  description: |-
    DataFusion SQL over the Iceberg tables, index and index-template management, batch jobs, GDPR delete tasks, and a Jaeger-compatible trace read API.

    Ingest lives on a separate server (`openapi-ingest.yaml`) — this one only reads, apart from index management and delete tasks.
  license:
    name: Apache-2.0
  version: 0.1.0
servers:
- url: http://localhost:8089
  description: Default `--bind` address
paths:
...
  /api/v1/sql/shard:
    post:
      tags:
      - sql
      summary: |-
        Worker endpoint: run the (sharded) query locally and return the result as an
        Arrow IPC stream (lossless transport). Not intended for human clients.
      operationId: shard
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ShardQueryRequest'
        required: true
      responses:
        '200':
          description: Partial result for this shard, as an Arrow IPC **stream** (lossless — the coordinator merges typed batches, not JSON).
          content:
            application/vnd.apache.arrow.stream:
              schema:
                type: array
                items:
                  type: integer
                  format: int32
                  minimum: 0
        '400':
          description: Empty or unparseable SQL, an invalid generation pin, a statement that is not a read (DDL, DML and session statements — `CREATE`, `COPY … TO`, `INSERT`, `SET` — are refused before they can take effect), or an invalid shard selector.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '401':
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '403':
          description: Tenant refused before the handler runs. This server derives the tenant from a verified JWT claim when one is configured (`--oidc-tenant-claim`), and this token carries none, or carries one that is not a usable tenant id (`[A-Za-z0-9_-]`, 1..=128 chars). Answered by the auth middleware, so it can reach every operation behind it.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '413':
          description: The mid-flight rows-scanned breaker tripped.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '503':
          description: 'Retryable refusal. Either this worker''s query memory pool (or spill directory byte cap) refused an allocation, or the request carried a `pin` whose generation this worker cannot resolve even after a metadata refresh (`reason: "shard_pin_unresolved"` in the body, with the pin) — a pinned shard is never answered from a different table incarnation or snapshot, because file shards only partition the table when every worker reads the same file generation, nor from a different schema, because a fan-out plans once on the coordinator. The coordinator forwards either to the caller rather than re-running the shard on itself.'
          headers:
            retry-after:
              schema:
                type: integer
                format: int64
                minimum: 0
              description: Seconds to wait before retrying.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
        '504':
          description: The shard exceeded its wall-clock budget. The budget starts when the request arrives and covers preparation (tenant resolution, table registration, pin resolution, planning) as well as execution, so a shard that spends it all preparing is refused without starting its scan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorBody'
...
components:
  schemas:
    ApiErrorBody:
      type: object
      description: |-
        Error body used by every endpoint on this server.

        Spec-only. Some errors additionally flatten diagnostic keys into the top
        level — `cost` on a rejected or timed-out query, and `cost` plus `stats` on
        the mid-flight rows breaker — so this schema deliberately allows additional
        properties rather than pinning an exact shape.
      required:
      - error
      - code
      properties:
        code:
          type: integer
          format: int32
          description: Echo of the HTTP status code.
          minimum: 0
        cost:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/CostReport'
            description: Present when the failure was cost- or time-related.
        error:
          type: string
          description: Human-readable failure description.
...
    ComplexityClass:
      type: string
      enum:
      - small
      - medium
      - large
      - huge
    CostReport:
      type: object
      description: |-
        Cost report returned by `/explain` and attached to every successful
        query response.
      required:
      - estimated_bytes_scanned
      - estimated_rows_processed
      - estimated_runtime_seconds
      - complexity_class
      - warnings
      - exact
      properties:
        complexity_class:
          $ref: '#/components/schemas/ComplexityClass'
        estimated_bytes_scanned:
          type: integer
          format: int64
          minimum: 0
        estimated_rows_processed:
          type: integer
          format: int64
          minimum: 0
        estimated_runtime_seconds:
          type: number
          format: double
        exact:
          type: boolean
          description: |-
            `true` when the underlying statistics are exact; `false` when
            they fell back to a heuristic.
        files_considered:
          type:
          - integer
          - 'null'
          description: |-
            Live files considered BEFORE the manifest time-bounds prune —
            `files_considered - files_to_scan` is planning-time pruning's win.
          minimum: 0
        files_to_scan:
          type:
          - integer
          - 'null'
          minimum: 0
        warnings:
          type: array
          items:
            type: string
...
    Priority:
      type: string
      description: Query priority — picks which tier's defaults + ceilings apply.
      enum:
      - interactive
      - batch
...
    ShardParam:
      type: object
      description: A `{index, count}` scan-shard selector on a query request.
      required:
      - index
      - count
      properties:
        count:
          type: integer
          minimum: 0
        index:
          type: integer
          minimum: 0
    ShardPin:
      type: object
      description: '#89/#1655: one table generation a shard read is pinned to.'
      required:
      - table
      properties:
        empty:
          type: boolean
          description: |-
            Pin the table's initial, snapshotless generation. This differs from
            omitting `pin`, which deliberately asks for the worker's current
            generation.
        schema_id:
          type:
          - integer
          - 'null'
          format: int32
          description: |-
            #2554: the Iceberg schema id the coordinator PLANNED against, from the
            same metadata generation as `snapshot_id`.

            A snapshot is not the whole identity of what a query is served from. An
            additive `migrate-schema` commits an `UpdateSchemaAction` and no data
            snapshot, so the schema id moves while the snapshot id stands still. A
            worker whose metadata cache predates that migration therefore RECOGNISES
            the pinned snapshot and answers from its narrow schema, and a fragment
            referencing the new column fails on that peer as invalid SQL while every
            other peer succeeds.

            A worker enforces this alongside the snapshot: it refreshes its metadata
            once, serves the pinned schema — including a historical one, so a worker
            AHEAD of the coordinator is not refused — and answers **503**
            (`reason: "shard_pin_unresolved"`) when the generation cannot be
            honoured at all.

            Omitted means the pre-#2554 contract: the snapshot alone is enforced and
            the worker plans against its own current schema. That is what an older
            COORDINATOR sends. In the other direction — a coordinator that sends it
            to an older WORKER — the field is ignored, so a mixed-version cluster
            keeps the old exposure until every replica is upgraded.
        snapshot_id:
          type:
          - integer
          - 'null'
          format: int64
          description: The exact committed snapshot to read. Mutually exclusive with `empty`.
        table:
          type: string
        table_uuid:
          type:
          - string
          - 'null'
          description: |-
            #2921: the Iceberg table incarnation the coordinator captured.
            Snapshot and schema identifiers can both repeat when a snapshotless
            managed index is dropped and recreated under the same name.

            A worker enforces this against the same cached table handle it uses to
            resolve the schema or snapshot. A mismatch refreshes once, then refuses
            the shard if the named incarnation is no longer available.

            Omitted means the pre-#2921 contract: snapshot and schema pins are
            enforced as supplied, without incarnation identity. This permits an
            older coordinator to address a newer worker. An older worker ignores a
            UUID sent by a newer coordinator, retaining the old exposure until all
            workers have been upgraded.
    ShardQueryRequest:
      type: object
      required:
      - query
      properties:
        pin:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ShardPin'
            description: |-
              #89/#1655: the exact table generation used to construct the
              coordinator's WAL partial — a snapshot by id (or the initial empty
              generation) and, since #2554, the schema id it planned against. The
              worker reads THAT generation so a fan-out is consistent cluster-wide: a
              worker whose cache runs ahead would otherwise double-serve rows the
              coordinator still counts in its buffer partial, or answer with a wider
              column set than the merge was planned for.

              A worker that cannot resolve the pinned generation (a snapshot expired,
              a schema unknown, or either still not visible after one metadata
              refresh) answers **503** rather than serving its own current one: file
              shards only partition the table when every worker enumerates the same
              file generation, and a two-phase merge only type-checks when every
              partial carries the schema the coordinator planned. Omit `pin` to read
              the worker's current generation deliberately.
        priority:
          $ref: '#/components/schemas/Priority'
          description: |-
            Original query priority, forwarded by the coordinator so the worker
            applies the matching tier's mid-flight breaker. Defaults to interactive
            (the only priority that is ever coordinated).
        query:
          type: string
        shard:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/ShardParam'
        tenant:
          type:
          - string
          - 'null'
          description: |-
            The tenant the ORIGINAL caller resolved to, forwarded by the coordinator.

            A shard request carries the coordinator's own credential, not the
            caller's, so without this the worker re-derives tenancy from a token that
            belongs to no tenant and reads the DEFAULT namespace for a tenant
            caller -- a silent cross-tenant read on the packaged 2-replica default.

            Honoured ONLY when the request authenticated with this node's own
            `coordinator_token` -- `CallerIdentity::coordinator`, decided once in the
            auth middleware. Otherwise a user who can reach `/shard` directly could
            name any tenant they liked.
...
  securitySchemes:
    query_token:
      type: http
      scheme: bearer
      description: '`Authorization: Bearer <token>`. With OIDC configured the token is a JWT and its claims select the tenant; otherwise it is matched against the cluster''s static token list. A server started with neither accepts unauthenticated requests.'
security:
- query_token: []
...